Pushing to Main (#7)
* Proper Fixed

Another branch screwed with me, so i deleted it. Those commits & this commit (made double jump in all platformers, made internet search bar work, improved volume visualizer) are here

* Update README.md

* More Stuff :D

Added a Super Auto Pets clone! Updated the platformers with: Dying, collectable currency, and running!

* Update README.md

* Small Update

Added A car driving thing and a enemy spawner to the endless shmup

* Few cool changes

Made a battle scene for super auto pets, Made upgrades for Idle game, and fixed a weird bug that happened with the window and control scene...

* Updated 3D Platformer

Made it so there is now a player, you can actually see what it is doing